Move more tests using passivation and EL into Full test group.
